PM-40520 - Restructure open-invite into module folder, drop macro change
Reorganize the open-organization-invite registration crossing from three
flat sibling files (open_org_invite_data.rs, seal_open_org_invite_data.rs,
unseal_open_org_invite_data.rs) into a single open_org_invite/ module:

- mod.rs       — versioned enum via generate_versioned_sealable!, re-exports
- wire_v1.rs   — RegistrationOpenOrgInviteDataV1 wire schema
- seal.rs      — public request/response types + seal path + shared helpers
- unseal.rs    — unseal path + framing split helper

With seal/unseal as children of open_org_invite instead of top-level
siblings, the versioned enum can stay module-private (visible to descendants)
and the macro no longer needs a $vis:vis prefix. Reverts the macro visibility
change from bd2af6977 in bitwarden-crypto; that capability wasn't mine to
add and the folder shape removes the need for it. Follow-up to the layout
alternative Copilot suggested on #1295 (relocate vs. teach the macro).

Renamed the seal-side public input `OpenOrgInviteData` to
`OpenOrgInviteSealRequest` per the SDK's [Noun][Op]Request convention in
CLAUDE.md. `SealedOpenOrgInvite` keeps its shape-neutral name and now
carries a doc explaining it is both the seal response and the unseal
request — either suffix would misrepresent one of the two roles.
