+++ crates/bitwarden-auth/src/registration/mod.rs
@@ -3,15 +3,14 @@
 //! mechanisms to establish their cryptographic state and register with
 //! the Bitwarden server
 
-mod open_org_invite_data;
+mod open_org_invite;
 mod post_keys_for_jit_password_registration;
 mod post_keys_for_key_connector_registration;
 mod post_keys_for_tde_registration;
 mod post_keys_for_user_password_registration;
 mod registration_client;
-mod seal_open_org_invite_data;
-mod unseal_open_org_invite_data;
 
+pub use open_org_invite::{OpenOrgInviteSealRequest, SealedOpenOrgInvite};
 pub use post_keys_for_jit_password_registration::{
     JitMasterPasswordRegistrationRequest, JitMasterPasswordRegistrationResponse,
 };
@@ -21,4 +20,3 @@ pub use post_keys_for_user_password_registration::{
     UserMasterPasswordRegistrationRequest, UserMasterPasswordRegistrationResponse,
 };
 pub use registration_client::{RegistrationClient, RegistrationError};
-pub use seal_open_org_invite_data::{OpenOrgInviteData, SealedOpenOrgInvite};
